datasheriff/requirements.txt
mcp>=1.0.0
anthropic>=0.40.0
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
slack-bolt>=1.18.0
pydantic>=2.0.0
